Addded LCD display
Added an LCD display using the Arduino LiquidCrystal library. Had to make some changes to the makefile to include the LiquidCrystal function, and a whole bunch of things it depends on. 

Also, as LiquidCrystal is C++, I redeclared main as main.cpp, to get access to the C++ code. I've also put the code that describes what to write into main, because I wanted to keep the number of C++ modules to a minimum. I may think of refactoring later.

Took out hard coded references to PORTD, because I couldn't see what they were for, and I was using PORTD for my LCD display.

On further testing, this works okay up to about 150 mm/min rate. Higher than that, the step generation seems to mess with talking to the LCD. Once motion stops, the display catches up, but during motion the display is pretty ugly.



Signed-off-by: = bob gates