Added backlash compensation
In the stepper.c file, I added backlash compensation to st_buffer_block: with each new block that is buffered, the code checks to see if any of the directions have changed, and if they have, a block is inserted that just moves in that direction with the backlash figure stored in the settings. The following block then picks up from where things left off. As start position is saved for the following block, the backlash block has no effect on actual position - which is what should happen: if backlash compensation is perfect, the postion the machine thinks that it is at is the position it will be at!