Added support for an I2C based HMI
Added I2C support to send position to another Arduino, and get button presses from it. The other Arduino puts position up on a display, and reads buttons from a PS2 controller. This routine uses the Wire library, but gives the grbl code a simple C interface to the HMI.