Fixed bug with backlash and queueing
The backlash compensation works by adding another command to the stepper buffer, as well as the command that does the next movement. The code was organised so that there should be one free space in the buffer before a command is placed on the buffer. However, if the command requires backlash compensation, grbl has to wait while a second slot becomes free. 

I've changed the buffer-full check to keep two blocks free, which has fixed the problem. There are also other small changes to the backlash compensation code to fix some potential bugs.

Signed-off-by: bobgates