Fixed race condition to do with arc drawing.
When drawing an arc, the arc drawing function checks to see if there is room in the queue, then adds the next segment. At the same time, grbl may get a query from outside as to whether the buffer is free, and may reply that it is, only for the free spot to be filled by the arc drawing. When the command arrives, it produces a buffer full error.

Now when grbl reports on whether the buffer has space, it also checks whether an arc is being drawn, and if so, reports the buffer is full. Once all the arc segments are on the buffer, it will only then report that there is free space on the buffer. 

There was a lot of fiddling in several files, but I think the only material change is in sp_quick_position, which reports st_buffer_full()||mc_in_arc(). I couldn't put the mc_in_arc command into st_buffer_full, because the mc_arc drawing code uses it to see if the buffer is empty.

Signed-off-by: bobgates