Output PWM on Arduino Pin 11 to reflect Spindle Speed commands S0 through S255.  S > 255 will set Pin 11 to digital high.  As a result, Timer 2 is exclusively used for PWM and Timer 0 overflow is used by stepper.c in its stead.  Also, currently the Z-axis Limit switch is disabled, though it could be moved to Arduino Pin A4/A5, etc.
Changes have been tested with a FlexMod P3 laser diode driver to control
laser intensity for use in burning shades of grays with a 2 watt laser
diode mounted on a Shapeoko 2.  Arduino Pin 12 is still used with M3/M4/M5
to allows complete shutoff of the laser, where S0 may leave the laser on
but not burning (to improve diode life).

Changes made and tested by Andrew L. Sandoval -
http://www.andrewlsandoval.com/